+++ src/syslog2irc/config.py
@@ -8,16 +8,20 @@ Configuration loading
 :License: MIT, see LICENSE for details.
 """
 
-from argparse import Namespace
+import dataclasses
 from dataclasses import dataclass
 from itertools import chain
 from pathlib import Path
-from typing import Any, Dict, Iterator, Set
+from typing import Any, Dict, Iterator, Optional, Set
 
 import rtoml
 
-from .irc import IrcChannel, IrcConfig
+from .irc import IrcChannel, IrcConfig, IrcServer
 from .router import Route
+
+
+DEFAULT_IRC_SERVER_PORT = 6667
+DEFAULT_IRC_REALNAME = 'syslog'
 
 
 @dataclass(frozen=True)
@@ -26,23 +30,54 @@ class Config:
     routes: Set[Route]
 
 
-def parse_config(
-    args: Namespace, routes_dict: Dict[int, Set[IrcChannel]]
-) -> Config:
+def parse_config(path: Path, routes_dict: Dict[int, Set[IrcChannel]]) -> Config:
     """Parse configuration."""
-    config_filename = args.config_filename
-    config_data = _load_config(config_filename)
+    irc_config = _load_config(path)
+    irc_config = _assemble_irc_config(irc_config, routes_dict)
 
-    irc_config = _assemble_irc_config(args, routes_dict)
     routes = _parse_routes(routes_dict)
 
     return Config(irc=irc_config, routes=routes)
 
 
-def _load_config(path: Path) -> Dict[str, Any]:
+def _load_config(path: Path) -> IrcConfig:
     """Load configuration from file."""
     data = rtoml.load(path)
-    return data
+
+    irc_config = _get_irc_config(data)
+
+    return irc_config
+
+
+def _get_irc_config(data: Dict[str, Any]) -> IrcConfig:
+    data_irc = data['irc']
+
+    server = _get_irc_server(data_irc)
+    nickname = data_irc['bot']['nickname']
+    realname = data_irc['bot'].get('realname', DEFAULT_IRC_REALNAME)
+    channels = set()
+
+    return IrcConfig(
+        server=server,
+        nickname=nickname,
+        realname=realname,
+        channels=channels,
+    )
+
+
+def _get_irc_server(data_irc: Any) -> Optional[IrcServer]:
+    data_server = data_irc.get('server')
+    if data_server is None:
+        return None
+
+    host = data_server.get('host')
+    if not host:
+        return None
+
+    port = int(data_server.get('port', DEFAULT_IRC_SERVER_PORT))
+    ssl = data_server.get('ssl', False)
+
+    return IrcServer(host=host, port=port, ssl=ssl)
 
 
 def _parse_routes(routes_dict: Dict[int, Set[IrcChannel]]) -> Set[Route]:
@@ -57,14 +92,8 @@ def _parse_routes(routes_dict: Dict[int, Set[IrcChannel]]) -> Set[Route]:
 
 
 def _assemble_irc_config(
-    args: Namespace, routes_dict: Dict[int, Set[IrcChannel]]
+    irc_config: IrcConfig, routes_dict: Dict[int, Set[IrcChannel]]
 ) -> IrcConfig:
-    """Assemble IRC configuration from command line arguments and routing config."""
+    """Complete IRC configuration with channels from routing config."""
     channels = set(chain(*routes_dict.values()))
-
-    return IrcConfig(
-        server=args.irc_server,
-        nickname=args.irc_nickname,
-        realname=args.irc_realname,
-        channels=channels,
-    )
+    return dataclasses.replace(irc_config, channels=channels)